Prepare for public release
Add README file describing the project.

Add copyright and license notice to each source file, and a copy of the
GPL v3 in the LICENSE file.

Make the project build with the current Go compiler.

Remove unneeded files, like `upload-scp.sh` which was used for uploading
to a BeagleBone black running Debian.

Add a copyright notice and GitHub link to the Web UI.

Remove `config.json.example`, since the daemon automatically creates it
if missing.